requirements.txt
pandas==1.5.3
pandas-ta==0.3.14b
scikit-learn==1.7.0
python-dotenv==1.1.1
oandapyV20==0.7.2
websocket-client>=1.6.1
requests>=2.31.0
python-dateutil>=2.8.2
tzlocal>=5.1
websockets>=12.0
colorama>=0.4.6
xgboost>=3.0.0